Fix command timeout when ignore_output is disabled and allow more than one wodle command

+++ src/wazuh_modules/wmodules.c
@@ -112,18 +112,20 @@ int wm_check() {
             next = j->next;
 
             if (i->context->name == j->context->name) {
-                mdebug1("Deleting repeated module '%s'.", j->context->name);
+                if(strcmp(j->context->name,"command")){
+                    mdebug1("Deleting repeated module '%s'.", j->context->name);
 
-                if (j->context->destroy)
-                    j->context->destroy(j->data);
+                    if (j->context->destroy)
+                        j->context->destroy(j->data);
 
-                if (j == wmodules) {
-                    wmodules = prev = next;
-                } else {
-                    prev->next = next;
+                    if (j == wmodules) {
+                        wmodules = prev = next;
+                    } else {
+                        prev->next = next;
+                    }
+
+                    free(j);
                 }
-
-                free(j);
             } else {
                 prev = j;
             }
